Core.Types

Documentation

data Bot

The Bot record type

Constructors

Bot

Fields

socket :: Handle

The socket that joebot should write to

config :: Config

The configuration set at startup

type Net = ReaderT Bot IO

The Net monad

data Msg

This is the message type for Plugin Processes

Constructors

Msg Text (Maybe Text) Text [Text]
Msg name channel cmdName args

Res [Text]
Res txts

Act Text
Act action_txt

Quit
Used for quitting applications

data Command

Constructors

Command

Fields

_cmdName :: Text

The name of the command

_arity :: Int

The arity of the command

_runCmd :: Text -> Maybe Text -> [Text] -> Net ()

The command given is either a sequential command | or a running thread in the background (i.e. has persistent state)

_help :: Text

what is shown if the command is invoked incorrectly

data Config

Constructors

Config

Fields

_nick :: Text

Nickname for the bot

_rname :: Text

Real name of the bot

_server :: String

Hostname of the IRC server the bot should connect to.

_port :: Int

Connection port. Defaults to 6667

_chan :: Text

Channel the bot should join

_pass :: Maybe Text

Password for NickServ if bot is on irc.freenode.net

_cmds :: [Command]

All the commands the bot can invoke

_jhooks :: [Text -> Text -> Net ()]

Join hooks

_phooks :: [Text -> Text -> Net ()]

Part/Quit hooks

data Response

Constructors

Req Request

Join Text Text

Part Text Text

Ping Text

Txt Text

data Request

Constructors

Request

Fields

_name :: Text

_chn :: Maybe Text

_cname :: Text

_tokens :: [Text]

runCmd :: Lens' Command (Text -> Maybe Text -> [Text] -> Net ())

help :: Lens' Command Text

cmdName :: Lens' Command Text

arity :: Lens' Command Int

server :: Lens' Config String

rname :: Lens' Config Text

port :: Lens' Config Int

phooks :: Lens' Config [Text -> Text -> Net ()]

pass :: Lens' Config (Maybe Text)

nick :: Lens' Config Text

jhooks :: Lens' Config [Text -> Text -> Net ()]

cmds :: Lens' Config [Command]

chan :: Lens' Config Text

tokens :: Lens' Request [Text]

name :: Lens' Request Text

cname :: Lens' Request Text

chn :: Lens' Request (Maybe Text)